Only install node_modules if they don't exist

+++ run.sh
@@ -1,10 +1,17 @@
 #!/bin/bash
 
 cd client
-npm install
-npm build
+if [ ! -d "./node_modules" ]; then
+  npm install
+fi
+npm run-script build
 
 cd ../server
-pip install --editable .
+if [ -x "$(command -v pip3)" ]; then # if pip3 executable exists in PATH
+  pip3 install --editable .
+else
+  pip install --editable .
+fi
 export FLASK_APP=server
+export FLASK_DEBUG=true
 flask run
